requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
memory-profiler==0.61.0
psutil==5.9.5
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
requests==2.31.0
telebot==0.0.5
urllib3==2.0.5
